pyproject.toml
[project]
name = "task-django-quotes"
version = "0.1.0"
description = ""
authors = [
    {name = "Vadim Medintsev",email = "m8adimka@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
